Added option to create stackable mocks
A new `-stackable` option in mockgen allows users to create mocks that
can be stacked on top of other mocks or another implementation. A
non-mock implementation can only be at the bottom of the stack.

This change is backward compatible. Mocks generated without the
`-stackable` option are identical to ones before this change.

see sample/stacked/... for examples on mocks can be stacked.